+++ src/all/background_page/error/passboltApiFetchError.js
@@ -1,0 +1,15 @@
+/**
+ * Network error
+ *
+ * @copyright (c) 2019 Passbolt SA
+ * @licence GNU Affero General Public License http://www.gnu.org/licenses/agpl-3.0.en.html
+ */
+
+class PassboltApiFetchError extends Error {
+  constructor(message, data) {
+    super(message);
+    this.data = data || {};
+  }
+}
+
+exports.PassboltApiFetchError = PassboltApiFetchError;